module_2_ai_agent/requirements.txt
crewai==0.80.0
google-cloud-aiplatform==1.87.0
langchain-google-vertexai==2.0.7
streamlit==1.41.0
python-dotenv==1.0.1
